chore: fix tests after refactoring
