+++ test/test_model.py
@@ -32,7 +32,8 @@ import ops
 import ops.testing
 from ops import pebble
 from ops._private import yaml
-from ops.jujucontext import JujuVersion, _JujuContext
+from ops.jujucontext import _JujuContext
+from ops.jujuversion import JujuVersion
 from ops.model import _ModelBackend
 from test.test_helpers import FakeScript
